Condensed banner slideshow to a slideshow folder within the general-components folder
