--- src/constants/banner-pics.js
@@ -1,131 +1,0 @@
-// Boarding
-import boardingBanner1 from '../images/boarding/boarding-slide-show-1.png'
-import boardingBanner2 from '../images/boarding/boarding-slide-show-2.png'
-import boardingBanner3 from '../images/boarding/boarding-slide-show-3.png'
-import boardingBanner5 from '../images/boarding/boarding-slide-show-5.png'
-import boardingBanner6 from '../images/boarding/boarding-slide-show-6.png'
-//Daycare
-import daycareBanner1 from '../images/daycare/daycare-slideshow-1.png'
-import daycareBanner2 from '../images/daycare/daycare-banner5.png'
-import daycareBanner3 from '../images/daycare/daycare-slideshow-3.png'
-import daycareBanner4 from '../images/daycare/daycare-slideshow-4.png'
-import daycareBanner5 from '../images/daycare/daycare-slideshow-5.png'
-//Grooming
-import groomingBanner1 from '../images/grooming/grooming-white-tan-dog.jpg'
-import groomingBanner2 from '../images/grooming/grooming-slideshow-2.png'
-import groomingBanner3 from '../images/grooming/grooming-slideshow-3.png'
-//Home:
-import stockDaycare from '../images/stock-minnie.jpeg'
-import stockBoard from '../images/stock-littlebit.jpeg'
-import test1 from '../images/pics/charlie-and-puppy.jpeg'
-import test2 from '../images/pics/black-dog-cocked-head.jpeg'
-import test3 from '../images/pics/daphne-and-another-outside.jpeg'
-import test4 from '../images/pics/employee-outside.jpeg'
-import test5 from '../images/pics/golden-w-black-dogs.jpeg'
-import test6 from '../images/pics/grey-dog-outside.jpeg'
-import test7 from '../images/pics/littles-aspen.jpeg'
-import test8 from '../images/pics/running-golden.jpeg'
-
-export const boardingImages = [
-    {
-        image: boardingBanner1,
-        alt: '',
-    },
-    {
-        image: boardingBanner2,
-        alt: ''
-    },
-    {
-        image: boardingBanner3,
-        alt: ''
-    },
-    {
-        image: boardingBanner5,
-        alt: ''
-    },
-    {
-        image: boardingBanner6,
-        alt: ''
-    }
-]
-
-export const daycareImages = [
-    {
-        image: daycareBanner4,
-        alt: ''
-    },
-    {
-        image: daycareBanner1,
-        alt: ''
-    },
-    {
-        image: daycareBanner2,
-        alt: ''
-    },
-    {
-        image: daycareBanner5,
-        alt: ''
-    }
-]
-
-export const groomingImages = [
-    {
-        image: groomingBanner1,
-        alt: ''
-    },
-    {
-        image: groomingBanner2,
-        alt: ''
-    },
-    {
-        image: groomingBanner3,
-        alt: ''
-    }
-]
-
-export const homeImages = [
-    {
-        image: test1,
-        alt: ''
-    },
-    {
-        image: test2,
-        alt: ""
-    },
-    {
-        image: test3,
-        alt: ""
-    },
-    {
-        image: test4,
-        alt: ''
-    },
-    {
-        image: test5,
-        alt: ''
-    },
-    {
-        image: test6,
-        alt: ""
-    },
-    {
-        image: test7,
-        alt: ""
-    },
-    {
-        image: test8,
-        alt: ''
-    },
-    {
-        image: daycareBanner3,
-        alt: ''
-    },
-    {
-        image: stockBoard,
-        alt: ""
-    },
-    {
-        image: stockDaycare,
-        alt: ""
-    }
-]
